+++ internal/http/client.go
@@ -350,11 +350,8 @@ func (h *Client) ParseResponse(res *http.Response) (resMsg types.Message, err er
 		var mediaType string
 		var params map[string]string
 		if len(contentType) > 0 {
-			if mediaType, params, err = mime.ParseMediaType(contentType); err != nil {
-				h.mErrRes.Incr(1)
-				h.mErr.Incr(1)
-				h.log.Errorf("Failed to parse media type: %v\n", err)
-				return
+			if mediaType, params, err = mime.ParseMediaType(contentType); err == nil {
+				h.log.Warnf("Failed to parse media type from Content-Type header: %v\n", err)
 			}
 		}