+++ public/service/example_cache_plugin_test.go
@@ -4,14 +4,13 @@ import (
 	"context"
 	"time"
 
-	"github.com/Jeffail/benthos/v3/public/x/service"
+	"github.com/Jeffail/benthos/v3/public/service"
 )
 
 // LossyCache is a terrible cache example and silently drops items when the
 // capacity is reached. It also doesn't respect TTLs.
 type LossyCache struct {
-	Capacity int `yaml:"capacity"`
-
+	capacity int
 	mDropped *service.MetricCounter
 	items    map[string][]byte
 }
@@ -24,7 +23,7 @@ func (l *LossyCache) Get(ctx context.Context, key string) ([]byte, error) {
 }
 
 func (l *LossyCache) Set(ctx context.Context, key string, value []byte, ttl *time.Duration) error {
-	if len(l.items) >= l.Capacity {
+	if len(l.items) >= l.capacity {
 		// Dropped, whoopsie!
 		l.mDropped.Incr(1)
 		return nil
@@ -37,7 +36,7 @@ func (l *LossyCache) Add(ctx context.Context, key string, value []byte, ttl *tim
 	if _, exists := l.items[key]; exists {
 		return service.ErrKeyAlreadyExists
 	}
-	if len(l.items) >= l.Capacity {
+	if len(l.items) >= l.capacity {
 		// Dropped, whoopsie!
 		l.mDropped.Incr(1)
 		return nil
@@ -59,21 +58,21 @@ func (l *LossyCache) Close(ctx context.Context) error {
 // implementation of the cache (type `LossyCache`) also contains fields that
 // should be parsed within the Benthos config.
 func Example_cachePlugin() {
-	configSpec, err := service.NewStructConfigSpec(func() interface{} {
-		return &LossyCache{
-			Capacity: 100,
-			items:    map[string][]byte{},
-		}
-	})
-	if err != nil {
-		panic(err)
-	}
+	configSpec := service.NewConfigSpec().
+		Summary("Creates a terrible cache with a fixed capacity.").
+		Field(service.NewIntField("capacity").Default(100))
 
-	err = service.RegisterCache("lossy", configSpec,
+	err := service.RegisterCache("lossy", configSpec,
 		func(conf *service.ParsedConfig, mgr *service.Resources) (service.Cache, error) {
-			c := conf.AsStruct().(*LossyCache)
-			c.mDropped = mgr.Metrics().NewCounter("dropped.just.cus")
-			return c, nil
+			capacity, err := conf.FieldInt("capacity")
+			if err != nil {
+				return nil, err
+			}
+			return &LossyCache{
+				capacity: capacity,
+				mDropped: mgr.Metrics().NewCounter("dropped_just_cus"),
+				items:    make(map[string][]byte, capacity),
+			}, nil
 		})
 	if err != nil {
 		panic(err)