+++ public/service/example_input_plugin_test.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"math/rand"
 
-	"github.com/Jeffail/benthos/v3/public/x/service"
+	"github.com/Jeffail/benthos/v3/public/service"
 )
 
 type GibberishInput struct {